requirements.txt
Flask
Flask-SQLAlchemy
Flask-Dance
psycopg2
gunicorn
boto3
flask-migrate
flask_cors
Flask-Limiter